Progress 05-01-2023
Fundamentally reworked the Haven rendering system so that it uses VT escape sequences solely. There is no more WindowsNative renderer or Cross-platform ConWrite renderer or anything of that sort.

Added debug method to App class which can dump a specified number of frame buffers to a file.

Added Unix P/Invoke for writing to stdout, expanded Windows P/Invoke for writing to stdout.

Reorganized file structure so that it is a bit more sensible.

TextBox is now entirely deleted in favor of ScrollableTextBox.

Added small optimization flag in the VTRenderContext class that determines whether or not to optimize away empty lines in any widget using VTDrawCharacterInfoBuffer (ie. ScrollableTextBox). This can yield some performance but also causes occluded widgets to render through the space where the buffer is rendered. Can be enabled if the user is sure that no widgets are occluded by the buffer.

(wow this is a very long commit)